add authentication to helios-master
Authentication is enabled in the master when the `--auth-enabled`
argument is present.

Authentication can be restricted to only apply to a certain client
version and above, to allow for rolling out the authentication
requirement slowly across a Helios installation. This is controlled with
the `--auth-minimum-version` argument when starting the master. If not
set, authentication is applied to all clients and requests.

Helios supports different types of authentication "schemes". This allows
for customization of the details of how to authenticate users in your
particular installation. Helios will ship with support for the
`crtauth` scheme, but additional schemes can be added with plugins
(more on this later). The scheme in use by the master is controlled
with the `--auth-scheme` argument when starting the master.

Loading the authentication scheme
====
Helios uses the `java.util.ServiceLoader` facility for loading
Authenticator instances from the classpath. The first Authenticator
instance found that matches the configured scheme name is used. If no
Authenticator instance is found, the master will fail to start up.

Additional JARs can be added to the lookup path (beyond the regular
CLASSPATH) with the `--auth-plugin` argument to the master.

Handling the authentication handshake
====
Implementations of Authenticator can specify an additional HTTP
endpoint / resource class that Helios will register with Jersey for
handling the actual authentication handshake, if needed. For
example, the `crtauth` scheme uses a challenge-response handshake
that requires clients to make two separate requests to the server in
order to obtain an access token.

Any endpoints registered by the Authenticator will be responsible
for setting up methods/resources within Jersey for mapping each step in
this handshake. The URI to the first step in the handshake is sent to
the client in the initial 401 Unauthorized response sent to indiciate to
the client that it needs to authenticate.

Authenticating requests
=====
When authentication is enabled, the master adds an
InjectableProvider to the Jersey context supplied by the plugin.
This InjectableProvider is responsible for examining HTTP
headers for credentials and verifying them.

When the Authorization header is not present or has invalid credentials,
and if authentication is enabled for the version that the client is
using, then a response of 401 Unauthorized is sent, containing the name
of the authentication scheme the client is to use in the
`WWW-Authenticate` header.

The client's authentication implementation is then expected to
communicate with the handshake endpoint to obtain an access token (if
necessary; some schemes may require no handshaking).