simplify how authentication hooks into Jersey
Instead of requiring that each individual resource method have a
parameter like `@Auth HeliosUser user` (even if the user info is not
accessed in that method), handle the authentication instead in a Jersey
filter that is automatically applied prior to invoking any resource
method.

This means we can skip having to add unused method parameters to all of
the Jersey resources, as well as the awkward `InjectableProvider`
implementation to hook into Jersey.

Note: the Jersey filter does not apply to anything accessing `/metrics`
on the admin port (since there is no Jersey servlets on that
port/connector).